refactor: Extract spending limit tx to the tx sender
