+++ src/services/tx/__tests__/spendingLimitParams.test.ts
@@ -6,7 +6,7 @@ import * as txSender from '@/services/tx/tx-sender/create'
 import * as spendingLimitParams from '@/services/tx/spendingLimitParams'
 import type Safe from '@safe-global/safe-core-sdk'
 import type { SpendingLimitState } from '@/store/spendingLimitsSlice'
-import { createNewSpendingLimitTx } from '@/components/settings/SpendingLimits/NewSpendingLimit/steps/ReviewSpendingLimit'
+import { createNewSpendingLimitTx } from '@/services/tx/tx-sender'
 
 const mockData: NewSpendingLimitData = {
   beneficiary: ZERO_ADDRESS,
