+++ src/services/tx/tx-sender/index.ts
@@ -1,2 +1,3 @@
 export * from './create'
 export * from './dispatch'
+export * from './spendingLimit'
